Fix an issue with --show-docstring
This commit fixes an issue that occurs when calling `help()` on passlib
methods in some versions.

I think this is a change in Python where `help()` tries to resolve
`classproperty` values on the base classes?  Not sure, but this used to
work in versions prior to Python 3.11.  For now, if `help()` fails, we
just try to print the method `__doc__` attribute in stead.